Moved create_sys_all_attrs to helperfunctions
