+++ vermouth/tests/helper_functions.py
@@ -17,8 +17,11 @@ Contains helper functions for tests.
 """
 import operator
 import os
+import networkx as nx
 import networkx.algorithms.isomorphism as iso
-
+import vermouth
+from vermouth.system import System
+from vermouth.forcefield import ForceField
 
 def make_into_set(iter_of_dict):
     """
@@ -79,3 +82,69 @@ def find_in_path(names=('martinize2', 'martinize2.py')):
             fullpath = os.path.join(folder, name)
             if os.path.isfile(fullpath):
                 return fullpath
+          
+def create_sys_all_attrs(molecule, moltype, secstruc, defaults, attrs):
+    """
+    Generate a test system from a molecule
+    with all attributes set and blocks in
+    force-field.
+
+    Parameters
+    ----------
+    molecule: :class:vermouth.Molecule
+    moltype: str
+        sets meta['moltype']
+    secstruc: dict[int, str]
+        secondary structure attributes
+        as resid str pairs
+    defaults: dict
+        dict of attribute default value
+    attrs: dict
+        dict of attribute name and dict
+        of node value pairs
+
+    Returns
+    -------
+    :class:vermouth.System
+    """
+    # set mol meta
+    molecule.meta['moltype'] = moltype
+    # assign default node attributes
+    for attr, value in defaults.items():
+        nx.set_node_attributes(molecule, value, attr)
+
+    # assign node attributes
+    for attr, values in attrs.items():
+        nx.set_node_attributes(molecule, values, attr)
+
+    # assign resids
+    resids = nx.get_node_attributes(molecule, "resid")
+    nx.set_node_attributes(molecule, resids, "_old_resid")
+   
+    # make the proper force-field
+    ff = ForceField("test")
+    ff.variables['water_type'] = "W"
+    ff.variables['regular'] = 0.47
+    ff.variables['small'] = 0.41
+    ff.variables['tiny'] = 0.38
+
+    res_graph = vermouth.graph_utils.make_residue_graph(molecule)
+    for node in res_graph.nodes:
+        mol_nodes = res_graph.nodes[node]['graph'].nodes
+        block = vermouth.molecule.Block()
+        resname = res_graph.nodes[node]['resname']
+        resid = res_graph.nodes[node]['resid']
+        # assign secondary structure
+        for mol_node in mol_nodes:
+            molecule.nodes[mol_node]['cgsecstruct'] = secstruc[resid]
+            block.add_node(molecule.nodes[mol_node]['atomname'],
+                           atype=molecule.nodes[mol_node]['atype'])
+
+        ff.blocks[resname] = block
+
+
+    # create the system
+    molecule._force_field = ff
+    system = System()
+    system.molecules.append(molecule)
+    return system
